+++ sample/src/main/java/de/pfaffenrodt/adapter/sample/presenter/SamplePresenterSelector.kt
@@ -7,6 +7,7 @@ import de.pfaffenrodt.adapter.sample.BR
 import de.pfaffenrodt.adapter.sample.SampleItemA
 import de.pfaffenrodt.adapter.sample.SampleItemB
 import de.pfaffenrodt.adapter.sample.SampleItemC
+import de.pfaffenrodt.adapter.sample.SampleItemD
 
 class SamplePresenterSelector: ClassPresenterSelector() {
     init {
@@ -21,6 +22,15 @@ class SamplePresenterSelector: ClassPresenterSelector() {
         addClassPresenter(
                 SampleItemC::class,
                 DataBindingPresenter(R.layout.item_sample_databinding, BR.item)
+        )
+        addClassPresenter(
+                SampleItemD::class,
+                SampleViewBindingPresenter {
+                    binding, item -> run {
+                        val sampleItem = item as SampleItemD
+                        binding.text.text = sampleItem.value
+                    }
+                },
         )
     }
 }